// src/app/auth/models/role.ts
export enum Role {
  Admin = 'Admin',
  Client = 'Client',
  User = 'User',
  PYT_21_Admin = 'Administrador',
  PYT_21_Inversor = 'Inversor',
  PYT_24_Admin = 'PYT_24_Admin',
  PYT_24_Minero = 'PYT_24_Minero',
  PYT_4_Director = 'Director',
  PYT_4_Gerente = 'Gerente',
  PYT_4_Logistica = 'Logistica',
  PYT_4_Chofer = 'Chofer',
  PYT_4_Cliente = 'Cliente',
  PM = 'Project Manager'
}

//he dejado los roles actuales para no generar conflictos por ahora pero los eliminaré